fix(super-editor): extract shared hash helpers
